@@ -21,7 +21,7 @@ void main() {
 }
 
 class AppleTVGestureHandler {
-  static AppleTVGestureHandler _instance;
+  static AppleTVGestureHandler? _instance;
 
   factory AppleTVGestureHandler() => _instance ??= AppleTVGestureHandler._();
 
@@ -32,8 +32,7 @@ class AppleTVGestureHandler {
   //static const keyEventChannelName = 'flutter/keyevent';
   static const codec = JSONMessageCodec();
 
-  static const channel =
-      BasicMessageChannel<dynamic>(gamePadChannelName, codec);
+  static const channel = BasicMessageChannel<dynamic>(gamePadChannelName, codec);
   //static const channel_key = BasicMessageChannel<dynamic>(keyEventChannelName, codec);
 
   AppleTVGestureHandler._();
@@ -43,7 +42,7 @@ class AppleTVGestureHandler {
 
     // (< flutter 1.26) Workarround for the fact that ios (and therefore also AppleTV) does not support keyboards and
     // therefor do not handle key events for ios in the Focusmanger in the flutter library (outside the scope of the engine)
-    RawKeyboard.instance.addListener((event) {
+    HardwareKeyboard.instance.addHandler((event) {
       if (event.runtimeType == RawKeyUpEvent) {
         if (LogicalKeyboardKey.arrowLeft == event.logicalKey) {
           _moveLeft();
@@ -59,12 +58,10 @@ class AppleTVGestureHandler {
           return true;
         } else if ((LogicalKeyboardKey.enter == event.logicalKey) ||
             (LogicalKeyboardKey.select == event.logicalKey)) {
-//          print("OK key up");
           return true;
         } else if ((LogicalKeyboardKey.backspace == event.logicalKey) ||
             (LogicalKeyboardKey.escape == event.logicalKey)) {
-          ByteData message = const JSONMethodCodec()
-              .encodeMethodCall(const MethodCall('popRoute'));
+          ByteData message = const JSONMethodCodec().encodeMethodCall(const MethodCall('popRoute'));
           ServicesBinding.instance.defaultBinaryMessenger
               .handlePlatformMessage('flutter/navigation', message, (_) {});
           return true;
@@ -72,57 +69,33 @@ class AppleTVGestureHandler {
       } else if (event.runtimeType == RawKeyDownEvent) {
         if ((LogicalKeyboardKey.enter == event.logicalKey) ||
             (LogicalKeyboardKey.select == event.logicalKey)) {
-//          print("OK key down");
           return true;
         }
       }
+      return false;
     });
   }
-
-/*
-  void _simulateSingleKeyEvent(int keyCode, String type, String keyMap) {
-    final data = codec.encodeMessage({
-      'keymap': 'keyMap',
-      'type': type,
-      'keyCode': keyCode,
-      'modifiers': 0,
-    });
-    channel_key.binaryMessenger.send(CHANNEL_NAME_KEYS, data);
-  }
-
-  void _simulateKeyEvent(int keyCode, String keyMap) {
-    _simulateSingleKeyEvent(keyCode, 'keydown', keyMap);
-    _simulateSingleKeyEvent(keyCode, 'keyup', keyMap);
-  }
-*/
 
 // Using "FocusManager.instance.primaryFocus.focusInDirection" is a workarround because ios target does not handle key presses in flutter focusmanager. Ideally we would send key events!
 // In the future when ios does support, the code needs to be update to simulate the key press. At that point also key codes sent need to be validated corrected in the code below!
 
   void _moveUp() {
-//    _simulateKeyEvent(0x00070052 /*0x7E*/, 'fusia');
-    FocusManager.instance.primaryFocus.focusInDirection(TraversalDirection.up);
+    FocusManager.instance.primaryFocus?.focusInDirection(TraversalDirection.up);
   }
 
   void _moveDown() {
-//    _simulateKeyEvent(0x100070051 /*0x7D*/, 'macos');
-    FocusManager.instance.primaryFocus
-        .focusInDirection(TraversalDirection.down);
+    FocusManager.instance.primaryFocus?.focusInDirection(TraversalDirection.down);
   }
 
   void _moveLeft() {
-//    _simulateKeyEvent(0x0007004f /*0x7B*/, 'macos');
-    FocusManager.instance.primaryFocus
-        .focusInDirection(TraversalDirection.left);
+    FocusManager.instance.primaryFocus?.focusInDirection(TraversalDirection.left);
   }
 
   void _moveRight() {
-//    _simulateKeyEvent(0x7C, 'macos');
-    FocusManager.instance.primaryFocus
-        .focusInDirection(TraversalDirection.right);
+    FocusManager.instance.primaryFocus?.focusInDirection(TraversalDirection.right);
   }
 
-  Future _onMessage(dynamic arguments) {
+  Future<void> _onMessage(dynamic arguments) async {
     num x = arguments['x'];
     num y = arguments['y'];
     String type = arguments['type'];
@@ -161,8 +134,6 @@ class AppleTVGestureHandler {
     } else if (type == 'ended') {
       isMoving = false;
     }
-
-    return null;
   }
 }
 
@@ -223,17 +194,17 @@ class _AppState extends State<App> {
 }
 
 class Tile extends StatefulWidget {
-  final double w;
-  final double h;
-  final Color color;
+  final double? w;
+  final double? h;
+  final Color? color;
   final bool autofocus;
 
   const Tile({
-    Key key,
+    Key? key,
     this.w,
     this.h,
     this.color,
-    this.autofocus,
+    this.autofocus = true,
   }) : super(key: key);
 
   @override
@@ -249,42 +220,38 @@ class _TileState extends State<Tile> {
     return Padding(
       padding: const EdgeInsets.all(8.0),
       child: Focus(
-        autofocus: widget.autofocus ?? true,
+        autofocus: widget.autofocus,
         onFocusChange: (value) => setState(() {
           focused = value;
 
           if (focused) {
-            _containerColor = Colors.blue[700];
+            _containerColor = Colors.blue;
             // do stuff
           } else {}
         }),
         child: Container(
           decoration: BoxDecoration(
             borderRadius: BorderRadius.circular(30.0),
-            color:
-                this.widget.color ?? (focused ? _containerColor : Colors.white),
+            color: this.widget.color ?? (focused ? _containerColor : Colors.white),
             image: DecorationImage(
               image: AssetImage('assets/images/flutter.png'),
             ),
           ),
-          width: focused ? widget.w : widget.w,
-          height: focused ? (widget.h) : widget.h,
+          width: widget.w,
+          height: widget.h,
         ),
-        onKey: (_, event) {
-          // if (!mounted) return false;
-
+        onKeyEvent: (_, event) {
           // In some cases RawKeyDownEvent isn't sent by engine.
           // To manage all cases use RawKeyUpEvent which is always sent.
           if (event is RawKeyUpEvent &&
               (event.logicalKey == LogicalKeyboardKey.enter ||
                   event.logicalKey == LogicalKeyboardKey.select)) {
-            //  widget._showVirtualKeyboardIfNeeded();
             setState(() {
-              _containerColor = Colors.red[400];
+              _containerColor = Colors.red;
             });
-            return true;
+            return KeyEventResult.handled;
           }
-          return false;
+          return KeyEventResult.ignored;
         },
       ),
     );
@@ -311,10 +278,10 @@ class Collection extends StatelessWidget {
 class AxisTraversalPolicy extends FocusTraversalPolicy {
   final Axis axis;
 
-  AxisTraversalPolicy({this.axis});
+  AxisTraversalPolicy({required this.axis});
 
   @override
-  FocusNode findFirstFocusInDirection(
+  FocusNode? findFirstFocusInDirection(
     FocusNode currentNode,
     TraversalDirection direction,
   ) {
@@ -322,7 +289,7 @@ class AxisTraversalPolicy extends FocusTraversalPolicy {
   }
 
   @override
-  FocusNode findFirstFocus(FocusNode fn) {
+  FocusNode? findFirstFocus(FocusNode fn) {
     return null;
   }
 
@@ -387,54 +354,41 @@ class AxisTraversalPolicy extends FocusTraversalPolicy {
   }
 
   void exitGroup(FocusNode node, TraversalDirection direction) {
-    var candidates = node.nearestScope.traversalDescendants.toList();
-    final newNode = _moveFocus(node, direction, candidates);
-
-    Scrollable.ensureVisible(
-      newNode.context,
-      alignment: 0.5,
-      alignmentPolicy: ScrollPositionAlignmentPolicy.explicit,
-      duration: const Duration(milliseconds: 250),
-    );
+    _ensureVisible(node.nearestScope?.traversalDescendants.toList(), node, direction);
   }
 
   void moveFocus(FocusNode node, TraversalDirection direction) {
-    var candidates = node.parent.traversalDescendants
-        .where((element) => element != node)
-        .toList();
-
-    final newNode = _moveFocus(node, direction, candidates);
-
-    Scrollable.ensureVisible(
-      newNode.context,
-      alignment: 0.5,
-      alignmentPolicy: ScrollPositionAlignmentPolicy.explicit,
-      duration: const Duration(milliseconds: 250),
-    );
+    _ensureVisible(node.parent?.traversalDescendants.where((element) => element != node).toList(),
+        node, direction);
   }
 
-  FocusNode _moveFocus(FocusNode node, TraversalDirection direction,
-      List<FocusNode> candidates) {
+  void _ensureVisible(List<FocusNode>? candidates, FocusNode node, TraversalDirection direction) {
+    if (candidates != null) {
+      final ctxt = _moveFocus(node, direction, candidates).context;
+      if (ctxt != null) {
+        Scrollable.ensureVisible(
+          ctxt,
+          alignment: 0.5,
+          alignmentPolicy: ScrollPositionAlignmentPolicy.explicit,
+          duration: const Duration(milliseconds: 250),
+        );
+      }
+    }
+  }
+
+  FocusNode _moveFocus(FocusNode node, TraversalDirection direction, List<FocusNode> candidates) {
     switch (direction) {
       case TraversalDirection.up:
-        candidates = candidates
-            .where((element) => element.rect.bottom < node.rect.top)
-            .toList();
+        candidates = candidates.where((element) => element.rect.bottom < node.rect.top).toList();
         break;
       case TraversalDirection.right:
-        candidates = candidates
-            .where((element) => element.rect.left > node.rect.right)
-            .toList();
+        candidates = candidates.where((element) => element.rect.left > node.rect.right).toList();
         break;
       case TraversalDirection.down:
-        candidates = candidates
-            .where((element) => element.rect.top > node.rect.bottom)
-            .toList();
+        candidates = candidates.where((element) => element.rect.top > node.rect.bottom).toList();
         break;
       case TraversalDirection.left:
-        candidates = candidates
-            .where((element) => element.rect.right < node.rect.left)
-            .toList();
+        candidates = candidates.where((element) => element.rect.right < node.rect.left).toList();
         break;
     }
 
@@ -446,27 +400,21 @@ class AxisTraversalPolicy extends FocusTraversalPolicy {
           case TraversalDirection.up:
             return (node.rect.topCenter - a.rect.bottomCenter)
                 .distance
-                .compareTo(
-                    (node.rect.topCenter - b.rect.bottomCenter).distance);
+                .compareTo((node.rect.topCenter - b.rect.bottomCenter).distance);
 
           case TraversalDirection.right:
             return (node.rect.centerRight - a.rect.centerLeft)
                 .distance
-                .compareTo(
-                    (node.rect.centerRight - b.rect.centerLeft).distance);
+                .compareTo((node.rect.centerRight - b.rect.centerLeft).distance);
           case TraversalDirection.down:
             return (node.rect.bottomCenter - a.rect.topCenter)
                 .distance
-                .compareTo(
-                    (node.rect.bottomCenter - b.rect.topCenter).distance);
+                .compareTo((node.rect.bottomCenter - b.rect.topCenter).distance);
           case TraversalDirection.left:
             return (node.rect.centerRight - a.rect.centerLeft)
                 .distance
-                .compareTo(
-                    (node.rect.centerRight - b.rect.centerLeft).distance);
+                .compareTo((node.rect.centerRight - b.rect.centerLeft).distance);
         }
-
-        return 0;
       });
 
     final newFocusNode = candidates.first;
@@ -476,7 +424,7 @@ class AxisTraversalPolicy extends FocusTraversalPolicy {
 
   @override
   bool inDirection(FocusNode currentNode, TraversalDirection direction) {
-    var items = sortDescendants(currentNode.parent.children, currentNode);
+    var items = sortDescendants(currentNode.parent?.children ?? [], currentNode);
 
     if (axis == Axis.vertical) {
       handleVerticalGroupNavigation(currentNode, direction, items);
@@ -488,8 +436,7 @@ class AxisTraversalPolicy extends FocusTraversalPolicy {
   }
 
   @override
-  Iterable<FocusNode> sortDescendants(
-      Iterable<FocusNode> descendants, FocusNode currentNode) {
+  Iterable<FocusNode> sortDescendants(Iterable<FocusNode> descendants, FocusNode currentNode) {
     int Function(FocusNode, FocusNode) compare;
 
     switch (axis) {
@@ -507,7 +454,6 @@ class AxisTraversalPolicy extends FocusTraversalPolicy {
 
 Future<bool> _onBackPressed() {
   exit(0);
-  return null;
 }
 
 class Home extends StatelessWidget {
@@ -520,8 +466,7 @@ class Home extends StatelessWidget {
             children: [
               Padding(
                   padding: EdgeInsets.fromLTRB(80, 40, 20, 20),
-                  child: Text('Flutter for Apple TV',
-                      style: TextStyle(fontSize: 28))),
+                  child: Text('Flutter for Apple TV', style: TextStyle(fontSize: 28))),
               Container(
                 child: ListView.builder(
                   itemCount: 4,
