@@ -18,8 +18,7 @@ void main() {
     // Verify that platform version is retrieved.
     expect(
       find.byWidgetPredicate(
-        (Widget widget) => widget is Text &&
-                           widget.data.startsWith('Running on:'),
+        (Widget widget) => widget is Text && (widget.data?.startsWith('Running on:') ?? false),
       ),
       findsOneWidget,
     );
